Update bootstrap fix for nav collapse
This is wired to the document, so you don't have to do it on ready() and it only gets called if the menu is already expanded. Some people have reported weird flashing where the menu opens and then immediately closes with other code that did not verify that the menu had the "in" class.